Refactor pick_new_version() to get_latest_tag_number()
It's useful to be able to determine the latest tag number without adding
one.  Therefore change the pick_new_version() function to simply find
the latest tag number and do not add one.

Also use "tag number" in the name instead of "version" since "number" is
also used as a variable.  I think this naming is clearer.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>