Add --edit option
It can be useful to work on the tag message before publishing a new
version.  This can be done using the new git publish --edit option.

Behind the scenes a -staging tag is created to store the unpublished tag
message.  The next time git publish is used the staging tag message will
be picked up.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>