Alias --message to --cover-letter
If you think about git-publish as a tagging tool then the concept of a
tag message makes sense.

If you think about git-publish as a email sending tool then the concept
of a cover letter makes sense.

They are both the same thing so alias the option for familiarity with
git-format-patch(1).

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>