Add quotes around arguments with spaces in DEBUG print
When DEBUG is True each git command invocation is printed, making it
easy to understand what commands are executed by the script.

Arguments that contain spaces should be quoted in the output so it's
clear that they are a single argv[] element.

This patch is not perfect because strings containing both spaces and
double-quotes will be formatted as "a b""def" or even "a" "bc".  But
that is probably a rare case which I don't want to overcomplicate with a
shell-correct quoting filter.

Reported-by: Jeff Nelson <jenelson@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>