Improve and default to --inspect-emails
Previously, we rely on git-send-email to prompt for confirmation before
sending patches. It has some disadvantages: there is no summary of the
whole series; when abort in the middle of the series, the first emails
are already out with no return.

The default mode (--inspect-emails) now runs git-send-email
non-interactively, after user confirmation about the summary.

The added mode (--no-inspect-emails) will directly send the mails
without asking.

Signed-off-by: Fam Zheng <famz@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>