Force final tag so that tags can be overridden
The --number option specifies the revision number to use for the patch.
If a tag was already created for the revision number the following error
is produced at the very end (after sending emails):

  fatal: tag 'topic-v1' already exists

Use the force flag on the final git tag so that the error does not
happen.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>