hook: Add pre-publish-tag hook
This allows running tests before writing cover letter. It may reveals
flaws in code and send the developer back to devel phase.

The one existing hook pre-publish-send-email is only invoked right
before sending the email which seems a bit too late for testing, and it
is not invoked unless --to or --cc is non-empty, therefore it is not
suitable as a point for starting quick test, expecially range compiling,
right after committing.

Signed-off-by: Fam Zheng <famz@redhat.com>