Ignore the default "." remote
Git branches created with "git checkout -b $BRANCH" command has a
default "." remote in .git/config. When sending pull requests in such a
branch, the per profile remote setting is masked.

Such a remote is obviously meaningless for pull requests, so simply
ignore it and fall through to the profile var if any.

Signed-off-by: Fam Zheng <famz@redhat.com>