Provide per-branch subject prefix
We have per-profile prefix and per-repo prefix, still not a per-branch
prefix. It'll be nice if we have that. For example, when we are working
on some RFC patches, we would like to have "RFC" prefix for a specific
branch, this is neither per-profile, nor per-repo.

Since we have "--subject-prefix", let's allow to store the value as
per-branch value.

Signed-off-by: Peter Xu <peterx@redhat.com>