git-publish: support worktrees when invoking hooks
In newer git versions there exist the concept of worktrees. With the
worktree feature was introduced a 'git rev-parse --git-common-dir' call,
this call points to the common git tree where the hooks directory
exists. The issue is rev-parse doesn't error out when given an unknown
option so to fallback for older git versions we test if rev-parse
returned the option we provided.

Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>