Use git to determine push remote when possible
Git has its own mechanisms for determining the default push location,
the nuances of which are not fully captured in our approximate function.

To this end, use the @{push} symbolic reference when possible to
determine the remote that git itself considers the default push
remote. This has been available in git since 2.5.

The approximate version will be kept as a fallback.