Fix UTF-8 output decoding from git(1)
The Popen(universal_newlines=True) argument was used to automatically
get decoded strings from Popen.  Unfortunately the default encoding in
Python 2 is 'ascii' so non-ASCII characters in commit descriptions
produce errors.

Drop universal_newlines so that Popen stdout and stderr are read as
binary.  Manually decode 'utf-8'.  In practice git logs are UTF-8
although this can be changed with the rarely used i18n.commitEncoding
and i18n.logOutputEncoding settings which are not supported by
git-publish.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>