Respect user/system encoding
It should not be assumed that commands will generate content encoded
in utf-8 on their stdout/stderr.  Unless it's stated that we'll only
interact with external utilities that produces utf-8 encoded content,
out best bet is to respect the user/system encoding.

Based on my experience, locale.getpreferredencoding() is the best
option for that.

Signed-off-by: Cleber Rosa <crosa@redhat.com>