Further standardize bytes/text handling
Commit 395e55bec standardized the handling of data comming from the
execution of commands and when writing to files.  But, there are other
sources of data, such as from command line options, that when not
given a proper encoding, are treated as bytes.

On Python 2, there's implicit conversion when mixing unicode and bytes
(and strings are bytes on Python 2).  When that happens, the content
of the bytes part will be decode using an 'ascii' codec.  A reproducer
is as simple as:

  >>> '\xc3\xa3' + u''

To avoid this implicit conversion, we need to manually standardize
the data into unicode.

Given that this is also intended to work on Python 3, a little type
check is necessary, given that something represented with quotes and
double-quotes in Python 3, is already an "unicode string" and thus,
cannot be decode()d.

An optional (probably desirable) next step is to standardize on
to_text(), instead of calling ".decode()" on the changes introduced by
395e55bec.

Finally, the reproducer a crash that this fixes is something like:

  $ git publish [...] --cc='Philippe Mathieu-Daudé <philmd@redhat.com>'

On a system using Python 2.

CC: Philippe Mathieu-Daudé <philmd@redhat.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>