Rename get_first_profile() to has_profiles()
get_first_profile() has a problem: it will give priority to
profiles in .gitpublish instead of .git/config.

The good news is that this doesn't matter: the only caller of
get_first_profile() doesn't care which profile is returned: it
only needs to know if any profile exists.

Rename the function and make it return a boolean, so nobody will
expect it to actually return the first profile.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>