Rewrite parse_header() for Python 2/3
The parse_header() function does not work with Python 3 because it mixes
str and bytes and uses the non-existent unicode().

Rewrite the function so it works for both Python 2 and 3.  Also drop the
catch-all exception handler that hides the details when something goes
wrong.  Exceptions are not expected in this code.

Reported-by: Fabiano Fidêncio <fabiano@fidencio.org>
Cc: Fam Zheng <fam@euphon.net>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>