Add parameter "--separate-send/-S"
Add a new parameter "--separate-send/-S" to decide whether to use a separate
git-send-email command to send each of the patches.  The default option is off,
which means all the patches will still be sent using one git-send-email process.

This new parameter can be used with --in-reply-to when there're more than one
patch to be sent as follow up of an existing series.  Without the new
parameter, --in-reply-to will send the extra patches as a standalone series (so
the 2nd/3rd/... patches will still reply to the 1st patch, rather than the
message specified in --in-reply-to).  The problem is the command line:

  git-send-email PATCH1 PATCH2 PATCH3 ...

Will always let PATCH2/PATCH3/... to reply to PATCH 1 even if --in-reply-to is
specified.

Assuming we've sent the cover letter and 2 patches already, but we still want
to continue send the patches 3-5.  Before this patch, the final patchset will
look like (using --in-reply-to after both old/new patchsets):

  old cover letter
    old patch 1
    old patch 2
    new patch 3
      new patch 4
      new patch 5

After this patch, we can use "--in-reply-to" plus "--separate-send", then it
will look like:

  old cover letter
    old patch 1
    old patch 2
    new patch 3
    new patch 4
    new patch 5

We can also use "--separate-send" upon a series that contains all the trivial
fixes, so the fixes will be sent one by one as single patch, rather than in a
whole series.

Signed-off-by: Peter Xu <peterx@redhat.com>