Avoid to encode the cover letter
git-send-email seems to be not happy with a cover letter already
encoded. In fact, git-format-patch generates it without encoding.

This problem occurs especially if 'sendemail.transferEncoding' is
not set to auto.

To avoid issues, instead of using set_content(), which encodes the
final result, we use set_payload() which doesn't encode.
git-send-email will encode the payload for us, respecting the
'sendemail.transferEncoding' setting.

Fixes: #88
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>