git-publish: use the git mbox format
There are many variations of the mbox format. Git uses '\n' line
endings, does not wrap lines, and has a From_ line.

The SMTP and HTTP policies that are currently used in the code do not
match git's mbox format precisely. Often this goes unnoticed because
there is some tolerance in git-send-email(1), but Peter Xu found that
'\r\n' line endings result in lines with trailing '\r' characters when
the quoted-printable transfer encoding is used by git-send-email(1).

Define git's mbox format policy and use it throughout the code. While
we're at it, make sure the From_ line is written when the cover letter
file is edited.

Originally-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>